feat: :recycle: move search from homepage to separate page
